% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelint.R
\name{levelint}
\alias{levelint}
\title{Check the Level of intergration and Check for Stationarity}
\usage{
levelint(timeseries, alpha = 0.05)
}
\arguments{
\item{timeseries}{A numeric vector or time series object.}

\item{alpha}{Significance level for the tests (e.g., 0.05).}
}
\value{
A differenced time series that has been made stationary.
}
\description{
This function makes a time series stationary by using the Augmented Dickey-Fuller (ADF),
Phillips-Perron (PP), and KPSS tests.
}
\examples{
library(tseries)
library(urca)
library(levelint)
data <- cumsum(rnorm(100)) # Random walk
levelint(data, alpha = 0.05)

}
